fix(nix): Workaround for injecting access token from agenix
Nix can't load tokens from a file, so instead of referencing the
secret containing the access token we reference a new secret that
contains a raw nix config file containing the token inline and
include it from the generated config file.

This works fine for system-wide tokens, but those secrets are not
readable from user invokations of `nix`, resulting in authorization
errors. To handle those the secret has to also be mounted as the user,
and re-included as part of the user config.

By default, Agenix exposes a path `"${XDG_CONFIG_HOME}/agenix/<secret name>"`
for user-level secrets. Nix configuration don't expand environment
variables, so we need to override the path with the `xdg.configHome`
known by home-manager instead of relying on the runtime env var.

- See https://github.com/NixOS/nix/issues/6536